Restructure sample app to allow easier additions of future samples
This makes the main activity a list of buttons that take you to different samples, which will make it easy to add new ones in the future. This also tweaks the UI a bit to look a little nicer
