+++ bottomsheet-sample/src/main/AndroidManifest.xml
@@ -6,7 +6,8 @@
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
-        android:theme="@style/AppTheme" >
+        android:theme="@style/AppTheme"
+        >
         <activity
             android:name=".sample.MainActivity"
             android:label="@string/app_name" >
@@ -15,6 +16,15 @@
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
+        </activity>
+        <activity
+            android:name=".sample.PickerActivity"
+            android:label="@string/intent_picker"
+            android:parentActivityName=".sample.MainActivity"
+            >
+            <meta-data
+                android:name="android.support.PARENT_ACTIVITY"
+                android:value=".sample.MainActivity" />
         </activity>
     </application>
