+++ bottomsheet-sample/src/main/res/values/strings.xml
@@ -1,4 +1,6 @@
 <resources>
-    <string name="app_name">BottomSheet-sample</string>
+    <string name="app_name">BottomSheet Sample</string>
     <string name="share">Share</string>
+    <string name="intent_picker">Intent Picker</string>
+    <string name="share_some_text">Share some text!</string>
 </resources>
